api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.9.10",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
line-ending = "lf"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # pep8-naming
    "N",
    # pylint basic checks
    "PL",
]
# Require key-value pairs in new line
extend-select = ["E231", "E241"]

# Ignore some common issues that might be too strict for a basic project
ignore = [
    "E501",  # Line too long (handled by formatter)
    "B905",  # `zip()` without an explicit `strict=` parameter
    "PLR2004",  # Magic numbers
]
# Allow automatically fixing issues
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["api"]
